extract stuff from edit.js and load on demand
